Modified step to respect version of numpy/scipy in requirements
The tarballs in npscipy-binaries are listed and if one of them
matches the version in requirements.txt, then that tarball is
used. Also, the tarballs are only cloned on an as-needed basis
to keep the slug size down.

If this is pulled or forked, you need to change lell to your github
name on bin/steps/npscipy line 13